// apps/git-client/src/generated/ShelfEntry.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RepositoryId } from "./RepositoryId";
import type { ShelfFile } from "./ShelfFile";

export type ShelfEntry = { id: string, repositoryId: RepositoryId, message: string, createdAtMs: number, files: Array<ShelfFile>, indexPatchChecksum: string, worktreePatchChecksum: string, };